requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy[asyncio]==2.0.30
asyncpg==0.28.0
alembic==1.13.1

# OpenAI
openai==1.30.1

# File uploads
python-multipart==0.0.9

# Config
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0
aiosqlite==0.20.0

# Utilities
aiofiles==23.2.1
